move Period to common
