+++ ebl/tests/corpus/test_chapter.py
@@ -1,5 +1,6 @@
 import attr
 import pytest
+from ebl.common.period import Period
 
 from ebl.corpus.domain.chapter import (
     Chapter,
@@ -13,7 +14,6 @@ from ebl.corpus.domain.line_variant import LineVariant
 from ebl.corpus.domain.manuscript import (
     Manuscript,
     ManuscriptType,
-    Period,
     PeriodModifier,
     Provenance,
     Siglum,
