+++ ebl/tests/corpus/test_manuscript.py
@@ -1,8 +1,9 @@
 from typing import Union
 
 import pytest
+from ebl.common.period import Period
 
-from ebl.corpus.domain.manuscript import Manuscript, ManuscriptType, Period, Provenance
+from ebl.corpus.domain.manuscript import Manuscript, ManuscriptType, Provenance
 from ebl.tests.factories.corpus import ManuscriptLineFactory
 from ebl.transliteration.domain.line import EmptyLine
 from ebl.transliteration.domain.line_number import LineNumber
