+++ ebl/tests/corpus/test_parse_chapter.py
@@ -1,6 +1,7 @@
 from typing import Sequence
 
 import pytest
+from ebl.common.period import Period
 
 from ebl.corpus.domain.line import Line
 from ebl.corpus.domain.manuscript_line import ManuscriptLine
@@ -8,7 +9,6 @@ from ebl.corpus.domain.line_variant import LineVariant
 from ebl.corpus.domain.manuscript import (
     Manuscript,
     ManuscriptType,
-    Period,
     Provenance,
     Siglum,
 )
